Moved out Analog, {UInt/Bool/SInt}FactoryBase to designated files
